# src/main/resources/openapi/viewer-api.yaml
...
openapi: 3.0.3
info:
  title: 'Tailormap API'
  description: '
  Tailormap API contains the calls necessary to get the configuration for a map viewer and calls
  to get features.
  
  Note: you need to send an X-XSRF-TOKEN header with the value from the XSRF-TOKEN
  cookie for non-GET requests. The Angular HttpClient does this automatically.
  '
  contact:
    name: 'B3Partners BV'
    url: 'https://www.b3partners.nl/'
    email: 'info@b3partners.nl'
  license:
    name: 'MIT'
    url: 'https://mit-license.org/'
  version: 'v1'
servers:
  - url: '{basePath}'
    variables:
      basePath:
        description: base path
        default: '/api'
components:
  securitySchemes:
    formAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
...
paths:
...
  /{viewerKind}/{name}/layer/{appLayerId}/proxy/{protocol}:
    parameters:
      - in: path
        name: viewerKind
        required: true
        schema:
          type: string
          enum:
            - app
            - service
      - description: 'viewer name'
        in: path
        name: name
        required: true
        schema:
          type: string
      - in: path
        name: appLayerId
        required: true
        schema:
          type: string
      - description: 'proxy protocol, must equal ''wmts'' or ''wms'''
        in: path
        name: protocol
        required: true
        schema:
          type: string
      - description: 'Must equal WMS or WMTS'
        in: query
        name: SERVICE
        required: true
        schema:
          type: string
      - name: REQUEST
        description: 'A valid WMS/WMTS request value.'
        in: query
        schema:
          type: string
        required: true
      - name: VERSION
        description: 'WMS/WMTS version.'
        in: query
        schema:
          type: string
        required: false
      - name: free-form
        description: Additional OGC and vendor-specific parameters.
        in: query
        schema:
          type: object
          additionalProperties: true
        style: form
        explode: true
        example:
          LAYER: mylayer
          BBOX: 1,2,3,4
    get:
      operationId: 'proxy'
      description: 'Proxy a WMS/WMTS request to a map service. Useful for accessing a map service requiring authentication
         without exposing the password to the frontend. Authentication is checked by Tailormap. Additional parameters 
         from the WMS spec may be specified which will be passed on to the original server.'
      security:
        - formAuth: [ ]
      responses:
        '200':
          description: 'OK'
        '400':
          description: 'Bad Request. May be returned for some combination of parameters that can not be processed or are incomplete.'
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: './status-responses.yaml#/components/schemas/RedirectResponse'
        '403':
          description: 'Forbidden. Proxy not enabled for service or the proxy has credentials but the app is public.'
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: 'Not Found. Application or application layer does not exist or protocol is unknown.'
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: 'Internal server error'
          content:
            text/plain:
              schema:
                type: string
        '504':
          description: 'Bad Gateway. Proxied service not reachable or responding.'
          content:
            text/plain:
              schema:
                type: string
    post:
      operationId: 'proxyPOST'
      security:
        - formAuth: [ ]
      description: 'The same as GET, but allows for a large request (for example with a large filter parameter) with
       (some) parameters in the body instead of the URL. The request is passed on to the original service as a GET 
       request. An XSRF token is required.'
      requestBody:
        description: Supply large parameters here to avoid extreme URL lengths.
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              description: OGC and vendor-specific parameters too large for the URL.
              type: object
              additionalProperties: true
              example:
                CQL_FILTER: INTERSECTS(geom, BUFFER(POINT(1 2), 10))
                SLD_BODY: ...
      responses:
        '200':
          description: 'OK'
        '400':
          description: 'Bad Request. May be returned for some combination of parameters that can not be processed or are incomplete.'
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: './status-responses.yaml#/components/schemas/RedirectResponse'
        '403':
          description: 'Forbidden. Proxy not enabled for service or the proxy has credentials but the app is public.'
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: 'Not Found. Application or application layer does not exist or protocol is unknown.'
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: 'Internal server error'
          content:
            text/plain:
              schema:
                type: string
        '504':
          description: 'Bad Gateway. Proxied service not reachable or responding.'
          content:
            text/plain:
              schema:
                type: string
...